# swagger.yaml
openapi: 3.0.3
info:
  title: ARMA Audit Engine API
  description: |
    GrowthScope / ARMA Audit Engine — generates local SEO audit reports for small businesses.

    **Primary endpoint (used by the dashboard UI):**
    - `POST /arma-report` — self-contained 5-page briefing vs a **manually supplied competitor**
      (`competitor_url`, optional `competitor_position`). One call, no prerequisites.

    **Legacy workflow (API-only, no longer in the UI):**
    1. `POST /lite-report` — scores a business, locks competitor data, returns a PDF (or JSON) lite report
    2. `POST /full-report` — deep-dive audit PDF; reads the competitor locked by step 1

    **Required env vars:** `DATAFORSEO_LOGIN` + `DATAFORSEO_PASSWORD` (primary map pack source),
    `GOOGLE_PLACES_API_KEY` (fallback), `ANTHROPIC_API_KEY` (analysis), `PAGESPEED_API_KEY` (scores).
  version: 1.0.0
  contact:
    email: pankajkumar.techie@gmail.com
servers:
  - url: https://www.arma-agency.site
    description: Production server
  - url: http://localhost:3002
    description: Local development server
...
paths:
...
  /lite-report:
    post:
      tags: [Reports]
      summary: Generate a Lite Audit Report
      description: |
        Fetches Google Business Profile data, map pack rankings (DataForSEO primary /
        Google Places fallback), monthly traffic, and generates a cold email + PDF lite report.

        **Must be run before `POST /full-report`** — persists competitor data to SQLite which
        the full report reads for consistency.

        Add `?format=json` (or `"format": "json"` in the body) to receive raw JSON instead of a PDF.
      parameters:
        - in: query
          name: format
          schema:
            type: string
            enum: [json]
          description: Return raw JSON instead of a PDF binary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LiteReportRequest'
            examples:
              pdf_default:
                summary: Roofer in Anchorage AK (returns PDF)
                value:
                  url: https://example-roofer.com
                  city: Anchorage
                  state: AK
                  vertical: Roofing
              json_format:
                summary: Dallas plumber — return JSON for debugging
                value:
                  url: https://example-plumber.com
                  city: Dallas
                  state: TX
                  vertical: Plumber
                  format: json
      responses:
        '200':
          description: |
            Default: `application/pdf` binary with
            `Content-Disposition: attachment; filename="ARMA_LiteCheck_<domain>.pdf"`.
            With `format=json`: JSON object (see schema).
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                $ref: '#/components/schemas/LiteReportResponse'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "url, city, state required"
        '502':
          description: Could not fetch map pack rankings from any configured data source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapPackError'
  /full-report:
    post:
      tags: [Reports]
      summary: Generate a Full Audit Report
      description: |
        Deep-dive audit PDF. **Requires `POST /lite-report` to have been called first** for
        the same domain — competitor data is read from SQLite for consistency.

        Runs:
        - Site crawl (Puppeteer) for lead and competitor
        - PageSpeed Insights (mobile + desktop) for both
        - Claude analysis (strengths, weaknesses, action plan)
        - Trust angle computation
        - Full-length branded PDF render
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FullReportRequest'
            examples:
              basic:
                summary: Full report for a domain with an existing lite report
                value:
                  url: https://example-roofer.com
      responses:
        '200':
          description: |
            Full audit PDF with
            `Content-Disposition: attachment; filename="ARMA_Audit_<domain>.pdf"`.
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Missing URL or no lite report found for this domain
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                no_url:
                  value: { error: "url required" }
                no_lite:
                  value: { error: "No Lite Report found. Call POST /lite-report first." }
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /arma-report:
    post:
      tags: [Reports]
      summary: Generate the ARMA Website Report (5-page editorial briefing)
      description: |
        Self-contained 5-page PDF that walks the customer journey gap-by-gap against a
        **manually supplied competitor** (`competitor_url`) — one page each for Trust
        (reviews vs competitor), Search (organic listing content), and Speed (mobile
        PageSpeed), each closing with an estimated monthly-loss banner, plus a cover
        and a closing "two ways forward" CTA.

        The competitor is no longer auto-selected by Google Maps/DataForSEO ranking,
        so the report carries no ranking or position data anywhere — Search only
        compares listing content (title/description), never rank numbers.

        Pulls live data in one call — Google Business Profile for both businesses,
        organic SERP listing content, mobile PageSpeed and a homepage reviews check —
        so no prior call is required.

        Add `?format=json` (or `"format": "json"` in the body) to receive the
        assembled report data as JSON instead of a PDF.
      parameters:
        - in: query
          name: format
          schema:
            type: string
            enum: [json]
          description: Return raw JSON instead of a PDF binary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArmaReportRequest'
            examples:
              pdf_default:
                summary: Plumber in Bronx NY vs a manually chosen competitor (returns PDF)
                value:
                  url: https://example-plumber.com
                  city: Bronx
                  state: NY
                  vertical: Plumbing
                  competitor_url: https://competitor.com
              pdf_with_competitor_position:
                summary: Junk removal in Plano TX with the competitor's known Google Maps rank
                value:
                  url: https://www.quickjunkremoval.org/
                  city: Plano
                  state: TX
                  vertical: Junk Removal
                  competitor_url: https://www.junk-king.com
                  competitor_position: 2
              json_format:
                summary: Return JSON for debugging
                value:
                  url: https://example-plumber.com
                  city: Bronx
                  state: NY
                  vertical: Plumbing
                  competitor_url: https://competitor.com
                  format: json
      responses:
        '200':
          description: |
            Default: `application/pdf` binary with
            `Content-Disposition: attachment; filename="ARMA_Report_<domain>.pdf"`.
            With `format=json`: the assembled report data object.
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                $ref: '#/components/schemas/ArmaReportData'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "url, city, state, competitor_url required"
        '500':
          description: Server error (upstream data source or PDF render failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Debug ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MapPackError:
      type: object
      required: [error]
      properties:
        error:
          type: string
        keywords_tried:
          type: array
          items:
            type: string
          description: Keywords the engine attempted before failing
        search_location:
          type: string
          description: City + state that was searched

    # ── Request bodies ───────────────────────────────────────────────────────────
    LiteReportRequest:
      type: object
      required: [url, city, state]
      properties:
        url:
          type: string
          format: uri
          description: Full URL of the lead's website
          example: https://example-roofer.com
        city:
          type: string
          description: City where the business operates
          example: Anchorage
        state:
          type: string
          description: Two-letter US state abbreviation
          example: AK
        vertical:
          type: string
          description: Business vertical / service category. Defaults to `"Home Services"` if omitted.
          example: Roofing
        format:
          type: string
          enum: [json]
          description: Return raw JSON instead of a PDF binary
    ArmaReportRequest:
      type: object
      required: [url, city, state, competitor_url]
      properties:
        url:
          type: string
          format: uri
          description: Full URL of the lead's website
          example: https://example-plumber.com
        city:
          type: string
          description: City where the business operates
          example: Bronx
        state:
          type: string
          description: Two-letter US state abbreviation
          example: NY
        vertical:
          type: string
          description: Business vertical / service category. Defaults to `"Home Services"` if omitted.
          example: Plumbing
        competitor_url:
          type: string
          format: uri
          description: Full URL of the manually chosen competitor's website — used for every comparison in the report instead of an auto-selected ranking competitor.
          example: https://competitor.com
        competitor_position:
          type: integer
          minimum: 1
          description: Optional manually supplied Google Maps rank for the competitor, shown as context only (e.g. "ranked #2 on Google Maps") on the cover page. Not fetched, not paired against a lead-side rank, and never used in the loss calculation.
          example: 2
        format:
          type: string
          enum: [json]
          description: Return raw JSON instead of a PDF binary
    ArmaReportData:
      type: object
      description: |
        The fully assembled report data returned when `format=json` — the exact object
        the PDF is rendered from.
      properties:
        date:
          type: string
          description: Report date, uppercase long form
          example: "JULY 10, 2026"
        city:
          type: string
          example: Plano
        state:
          type: string
          example: TX
        vertical:
          type: string
          example: Junk Removal
        organicQuery:
          type: string
          description: Buyer-intent search phrase used for the organic SERP snapshot
          example: junk removal Plano
        lead:
          type: object
          description: The audited business
          properties:
            name:
              type: string
              example: Quick Junk Removal
            rating:
              type: number
              format: float
              description: Google Business Profile star rating
              example: 5
            reviews:
              type: integer
              description: Google review count
              example: 255
            domain:
              type: string
              example: quickjunkremoval.org
            speed:
              type: integer
              nullable: true
              description: Mobile PageSpeed score 0–100, `null` when the check failed
              example: 62
            reviewsShownOnHome:
              type: integer
              description: Approximate number of reviews surfaced on the homepage (0 = none found)
              example: 0
        competitor:
          type: object
          description: The manually supplied competitor
          properties:
            name:
              type: string
              example: Junk King Plano
            rating:
              type: number
              format: float
              example: 4.9
            reviews:
              type: integer
              example: 491
            domain:
              type: string
              example: junk-king.com
            speed:
              type: integer
              nullable: true
              example: 97
            title:
              type: string
              description: Competitor's organic listing title ("" when not found)
              example: "Junk Removal Services in Plano, TX"
            description:
              type: string
              description: Competitor's organic listing description ("" when not found)
            position:
              type: integer
              nullable: true
              description: The manually supplied `competitor_position`, echoed back — context only
              example: 2
        isUrgent:
          type: boolean
          description: Whether the vertical is emergency-driven (affects report copy)
        organicAvailable:
          type: boolean
          description: Whether an organic SERP snapshot could be fetched
        leadAppearsOrganically:
          type: boolean
        competitorAppearsOrganically:
          type: boolean
        activeGaps:
          type: object
          description: Which measurable steps are genuine gaps for this business
          properties:
            trust:
              type: boolean
            speed:
              type: boolean
        totalLoss:
          type: array
          description: Estimated total monthly loss range in USD `[low, high]`
          items:
            type: number
          minItems: 2
          maxItems: 2
          example: [343, 636]
        gaps:
          type: object
          description: Per-gap monthly loss ranges in USD `[low, high]`
          properties:
            trust:
              type: array
              items:
                type: number
              example: [171, 318]
            speed:
              type: array
              items:
                type: number
              example: [171, 318]
    FullReportRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: Full URL of the lead's website (domain must match an existing lite report in SQLite)
          example: https://example-roofer.com

    # ── Map pack ─────────────────────────────────────────────────────────────────
...
    RankingKeyword:
      type: object
      properties:
        keyword:
          type: string
        position:
          type: integer
    RevenueData:
      type: object
      properties:
        monthly_loss:
          type: integer
          description: Estimated monthly revenue loss from missed rankings (USD)
        annual_loss:
          type: integer
          description: Estimated annual revenue loss (USD)
    ColdEmail:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
    ReviewInsights:
      type: object
      nullable: true
      properties:
        repliedCount:
          type: integer
        unansweredCount:
          type: integer
        totalChecked:
          type: integer
        replyRate:
          type: number
          format: float
        replyDataAvailable:
          type: boolean
        snippets:
          type: array
          items:
            type: string
          maxItems: 3
    LeadData:
      type: object
      properties:
        name:
          type: string
        rating:
          type: number
          format: float
        review_count:
          type: integer
        place_id:
          type: string
          nullable: true
        gbp_url:
          type: string
          format: uri
          nullable: true
        position:
          type: integer
          description: Weighted Google Maps pack position
        organic_position:
          type: integer
          nullable: true
          description: Organic SERP position for the primary keyword
        position_by_keyword:
          type: array
          items:
            $ref: '#/components/schemas/RankingKeyword'
        phone:
          type: string
        address:
          type: string
        owner:
          type: string
          description: Owner name extracted from the homepage by Claude
        service_area:
          type: string
          description: Service area extracted from the homepage by Claude
    CompetitorData:
      type: object
      properties:
        name:
          type: string
        rating:
          type: number
          format: float
        review_count:
          type: integer
        position:
          type: integer
        domain:
          type: string
          nullable: true
        place_id:
          type: string
          nullable: true
        phone:
          type: string
        owner:
          type: string
        service_area:
          type: string
        gbp_url:
          type: string
          format: uri
          nullable: true
    LiteReportResponse:
      type: object
      description: Returned only when `format=json` is passed
      properties:
        domain:
          type: string
        city:
          type: string
        state:
          type: string
        vertical:
          type: string
        niche_matched:
          type: string
          description: Canonical niche key matched to internal benchmark data
        search_location:
          type: string
          example: "Anchorage, AK"
        lead:
          $ref: '#/components/schemas/LeadData'
        competitor:
          $ref: '#/components/schemas/CompetitorData'
        fullPack:
          type: array
          items:
            type: object
          description: All businesses returned in the map pack query
        traffic_monthly:
          type: integer
          nullable: true
          description: Estimated monthly organic traffic (SimilarWeb / SemRush)
        revenue:
          $ref: '#/components/schemas/RevenueData'
        review_insights:
          $ref: '#/components/schemas/ReviewInsights'
        position_data_source:
          type: string
          description: Which source provided the map pack data
        ranking_method:
          type: string
          example: exact_google_maps
        ranking_keywords:
          type: array
          items:
            $ref: '#/components/schemas/RankingKeyword'
        gap_summary:
          type: string
          description: One-line summary of the rank gap vs the top competitor
        cold_email:
          $ref: '#/components/schemas/ColdEmail'